+++ src/main/java/dandi/dandi/pushnotification/adapter/in/web/dto/PushNotificationAllowanceUpdateRequest.java
@@ -1,8 +1,7 @@
-package dandi.dandi.pushnotification.web.dto;
+package dandi.dandi.pushnotification.adapter.in.web.dto;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import dandi.dandi.pushnotification.application.port.in.PushNotificationAllowanceUpdateCommand;
-import io.swagger.v3.oas.annotations.media.Schema;
 
 public class PushNotificationAllowanceUpdateRequest {
